Build Warning 제거.

+++ DX12App/InGameUI.h
@@ -81,7 +81,7 @@ public:
     void ShowGoAnim(); 
 
     std::mutex& GetMutex() { return mScoreboardMutex; }
-    float GetSpeed() const { return mCurrentSpeed; }
+    float GetSpeed() const { return static_cast<float>(mCurrentSpeed); }
 
     void SetTimeMinSec(int& m, int& s);
     //void SetWarningText() { GetTextBlock()[7].strText.assign("WARNING"); }
